requirements/base.txt
Django==3.0.4
django-environ==0.4.5
django-filter==2.2.0
djangorestframework==3.11.0
psycopg2==2.7.5
sentry-sdk==0.14.1